chg: dev: Modified decorator of iperf libraries
Modified the number of state classes to just 1. Testing and documentation confirm only 1 decorator is allowed per library on the current topology implementation.
Currently, if iperf is run as client and then run a server function for example on the same device; the function fails because the decorator is set for the client state, not the server one.